fix: move helpers to a shared internal package, plan to remove
